Read through actual p467-468
Added a new class "Box" - When the user touches BoxDrawingView, a new Box will be created and added to an array of existing boxes

Back in BoxDrawingView, add some code that uses your new Box object to track your drawing state.
Any time an ACTION_DOWN motion event is received, you set mCurrentBox to be a new Box with its origin as the event’s location. This new Box is added to the array of boxes. (In the next section, when you implement custom drawing, BoxDrawingView will draw every Box within this array to the screen.)
As the user’s finger moves around the screen, you update mCurrentBox.mCurrent. Then, when the touch is canceled or when the user’s finger leaves the screen, you null out mCurrentBox to end your draw motion. The Box is complete; it is stored safely in the array but will no longer be updated about motion events.

Notice the call to invalidate() in the case of ACTION_MOVE. This forces BoxDrawingView to redraw itself so that the user can see the box while dragging across the screen. Which brings you to the next step: drawing the boxes to the screen.